feat: SMS_PROVIDER feature flag
