+++ .env.example
@@ -65,6 +65,10 @@ NEXT_PUBLIC_LOG_LEVEL="info" # debug | info | warn | error | custom
 
 SUPPRESS_SENTRY_ERRORS_ON_LOCAL="true"
 
+# Whether we use Capitol Canary or Twilio to send SMS
+# capitol-canary | twilio
+SMS_PROVIDER="capitol-canary"
+
 # Whether or not to log the sql queries coming from Prisma
 LOG_DATABASE="false"
